# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Trainerplan API
  version: 1.0.0
  description: |
    Versionierte HTTP-API für die Trainerplan-Web-App und native Clients.
    Browser verwenden eine HTTP-only-Cookie-Session; native Clients verwenden
    das beim Login ausgegebene Bearer-Token.
servers:
  - url: /api/v1
    description: Aktuelle API-Version
...
security:
  - bearerAuth: []
  - sessionCookie: []
paths:
...
  /onboarding/club:
    post:
      tags: [Authentifizierung]
      summary: Verein nach der ersten Anmeldung einrichten
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, clubName, ageGroup]
              properties:
                name: { type: string, minLength: 2, maxLength: 100 }
                clubName: { type: string, minLength: 2, maxLength: 120 }
                designation: { type: string, maxLength: 80, description: Optionale Zusatzbezeichnung, z. B. Blau oder Mädchen }
                ageGroup: { type: string, enum: [g2, g1, f2, f1, e2, e1, d2, d1, c2, c1, b2, b1, a2, a1] }
      responses:
        '200': { description: Verein eingerichtet }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /exercises:
    get:
      tags: [Übungen]
      summary: Übungsbibliothek laden
      responses:
        '200': { description: Übungsbibliothek, content: { application/json: { schema: { type: object, properties: { exercises: { type: array, items: { $ref: '#/components/schemas/Exercise' } } } } } } }
    post:
      tags: [Übungen]
      summary: Übung erstellen
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Exercise' } } }
      responses:
        '201': { description: Übung erstellt, content: { application/json: { schema: { type: object, properties: { exercise: { $ref: '#/components/schemas/Exercise' } } } } } }
  /exercises/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceId' }]
    patch:
      tags: [Übungen]
      summary: Übung ändern
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Exercise' } } }
      responses:
        '200': { description: Übung geändert, content: { application/json: { schema: { type: object, properties: { exercise: { $ref: '#/components/schemas/Exercise' } } } } } }
    delete:
      tags: [Übungen]
      summary: Übung löschen
      responses: { '200': { $ref: '#/components/responses/Ok' } }
...
  /settings:
    get:
      tags: [Mannschaft]
      summary: Vereinseinstellungen laden
      responses: { '200': { description: Einstellungen geladen, content: { application/json: { schema: { type: object, properties: { settings: { $ref: '#/components/schemas/Settings' } } } } } } }
    put:
      tags: [Mannschaft]
      summary: Vereinseinstellungen speichern
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Settings' } } }
      responses: { '200': { description: Einstellungen gespeichert, content: { application/json: { schema: { type: object, properties: { settings: { $ref: '#/components/schemas/Settings' } } } } } } }
...
  /invitations:
    post:
      tags: [Einladungen]
      summary: Einladung erstellen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email: { type: string, format: email }
                name: { type: string }
                role: { $ref: '#/components/schemas/Role' }
                groupId: { type: [string, 'null'] }
                sendEmail: { type: boolean }
      responses: { '201': { description: Einladung erstellt, content: { application/json: { schema: { type: object, properties: { invitation: { $ref: '#/components/schemas/Invitation' }, link: { type: string, format: uri }, emailSent: { type: boolean } } } } } } }
  /invitations/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceId' }]
    delete:
      tags: [Einladungen]
      summary: Einladung zurückziehen
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /invitations/accept:
    get:
      tags: [Einladungen]
      security: []
      summary: Einladung prüfen
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses: { '200': { description: Einladung ist gültig }, '404': { $ref: '#/components/responses/NotFound' } }
    post:
      tags: [Einladungen]
      security: []
      summary: Einladung annehmen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                name: { type: string }
                password: { type: string, minLength: 12, maxLength: 256 }
      responses: { '200': { description: Zugang erstellt }, '409': { $ref: '#/components/responses/Conflict' } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: session-token }
    sessionCookie: { type: apiKey, in: cookie, name: trainerplan_session }
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 160 }
  responses:
    Ok:
      description: Erfolgreich
      content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } }
...
    Unauthorized:
      description: Nicht angemeldet
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    NotFound:
      description: Ressource nicht gefunden
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Konflikt
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties: { error: { type: string } }
    Ok:
      type: object
      properties: { ok: { type: boolean, example: true } }
    Role: { type: string, enum: [admin, trainer, player] }
...
    Exercise:
      type: object
      required: [id, title, description, duration, players, ageGroup, ageRange, category, accent, intensity, focus, setup, coaching, materials, fieldSize, variant]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        duration: { type: integer, minimum: 1, maximum: 300 }
        players: { type: string }
        ageGroup: { type: string, enum: [F-Jugend] }
        ageRange: { type: string, enum: [U8, U9, U8/U9] }
        category: { type: string, enum: [Ankommen, Einstieg, Hauptteil, Abschlussspiel] }
        accent: { type: string, pattern: '^#[0-9a-fA-F]{6}$' }
        intensity: { type: string, enum: [Niedrig, Mittel, Hoch] }
        focus: { type: array, items: { type: string } }
        setup: { type: string }
        coaching: { type: array, items: { type: string } }
        materials: { type: array, items: { type: object, required: [id, count], properties: { id: { type: string }, count: { type: integer, minimum: 1 } } } }
        fieldSize: { type: string }
        variant: { type: integer, minimum: 0 }
        youtubeUrl: { type: string, format: uri }
        trainerId: { type: [string, 'null'] }
        internalTeam: { type: [string, 'null'], enum: [A, B, null] }
...
    Settings:
      type: object
      required: [theme, teamFeatureEnabled, attendanceEnabled, waitlistEnabled, showResponsesToPlayers, automaticReminders, splitTeamsEnabled, trainingDeadlineHours, tournamentDeadlineHours, eventDeadlineHours, defaultTrainingCapacity, defaultTournamentCapacity, clubName, teamName, tournamentMinFYouth, tournamentMaxTeamSize, tournamentTrainerRequired, tournamentNotifications, tournamentDefaultSquadName, ageGroupIds]
      properties:
        theme: { type: string, enum: [dark, light] }
        teamFeatureEnabled: { type: boolean }
        attendanceEnabled: { type: boolean }
        waitlistEnabled: { type: boolean }
        showResponsesToPlayers: { type: boolean }
        automaticReminders: { type: boolean }
        splitTeamsEnabled: { type: boolean }
        trainingDeadlineHours: { type: integer, minimum: 0 }
        tournamentDeadlineHours: { type: integer, minimum: 0 }
        eventDeadlineHours: { type: integer, minimum: 0 }
        defaultTrainingCapacity: { type: integer, minimum: 1 }
        defaultTournamentCapacity: { type: integer, minimum: 1 }
        clubName: { type: string }
        teamName: { type: string }
        tournamentMinFYouth: { type: integer, minimum: 0 }
        tournamentMaxTeamSize: { type: integer, minimum: 0 }
        tournamentTrainerRequired: { type: boolean }
        tournamentNotifications: { type: boolean }
        tournamentDefaultSquadName: { type: string }
        ageGroupIds: { type: array, items: { type: string } }
...
    Invitation:
      type: object
      required: [id, email, name, role, ageGroup, expiresAt, createdAt]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        ageGroup: { type: string }
        groupId: { type: [string, 'null'] }
        invitedBy: { type: string }
        expiresAt: { type: string, format: date-time }
        acceptedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
...