# src/app/api/v1/settings/route.ts
import { NextRequest, NextResponse } from "next/server";
import { authenticatedUser } from "@/lib/auth";
import { apiError, readJson } from "@/lib/api-security";
import { getSettings, saveSettings } from "@/lib/settings";

export const dynamic = "force-dynamic";

export async function GET(request: NextRequest) {
  const user = await authenticatedUser(request);
  if (!user) return NextResponse.json({ error: "Nicht angemeldet." }, { status: 401 });
  return NextResponse.json({ settings: await getSettings(user) });
}

export async function PUT(request: NextRequest) {
  const user = await authenticatedUser(request);
  if (!user) return NextResponse.json({ error: "Nicht angemeldet." }, { status: 401 });
  if (user.role !== "admin") return NextResponse.json({ error: "Nur Admins dürfen Einstellungen ändern." }, { status: 403 });
  try {
    return NextResponse.json({ settings: await saveSettings(await readJson(request, 256_000), user) });
  } catch (error) {
    const result = apiError(error);
    return NextResponse.json({ error: result.message }, { status: result.status });
  }
}